pyproject.toml
[tool.poetry]
name = "document-chatbot"
version = "0.1.0"
description = "A CLI tool for document Q&A using LLMs and vector search"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "backend_api", from = "backend/src" },
    # { include = "cli/src" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.115.9"
uvicorn = { extras = ["standard"], version = "0.34.2" }
python-multipart = "0.0.20"
chromadb = "1.0.9"
langchain = "0.3.25"
langchain-community = "0.3.24"
openai = "1.79.0"
PyMuPDF = "1.25.5"
google-cloud-aiplatform = "1.93.0"
python-dotenv = "1.1.0"
pydantic = "2.11.4"
sse-starlette = "2.3.5"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
black = "25.1.0"
mypy = "1.15.0"
ruff = "0.11.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chatbot = "cli.src.commands.main:app"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"
